# Conflicts:
#	Pod/Assets/SCStoryboard.storyboard
#	Pod/Classes/SCBoardController.m
#	Pod/Classes/SCBubbleViewIn.h
#	Pod/Classes/SCBubbleViewOut.h
#	Pod/Classes/SCChatController.h
#	Pod/Classes/SCChatController.m
#	Pod/Classes/SCComposeMessageController.m
#	Pod/Classes/SCDialButton.h
#	Pod/Classes/SCFriendDetailController.m
#	Pod/Classes/SCFriendListController.m
#	Pod/Classes/SCGroupDetailController.m
#	Pod/Classes/SCGroupDetailHeaderController.m
#	Pod/Classes/SCHorizontalLineView.h
#	Pod/Classes/SCLoginController.m
#	Pod/Classes/SCRegistrationController.m
#	Pod/Classes/SCWaitIndicatorController.m

+++ Pod/Classes/SCLoginController.m
@@ -61,8 +61,8 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
-	int row = indexPath.row;
-	int section = indexPath.section;
+	int row = (int)indexPath.row;
+	int section = (int)indexPath.section;
 	switch (section) {
 		case 0:
 			if (row == 0)
